feat: put the request readings handler logic into a separate hook
