+++ src/Hooks/index.ts
@@ -12,3 +12,4 @@ export { default as useRewardPopup } from "./useRewardPopup/useRewardPopup";
 export { default as useInfoPopup } from "./useInfoPopup/useInfoPopup";
 export { default as useErrorPopup } from "./useErrorPopup/useErrorPopup";
 export { default as useSettingsButton } from "./useSettingsButton/useSettingsButton";
+export { default as useReadings } from "./useReadings/useReadings";
