+++ src/Hooks/useReadings/types.ts
@@ -1,0 +1,8 @@
+import { Path } from "@/types";
+
+export type UseReadingsPropTypes = {
+  cardsQty: number;
+  path: Path;
+  prompt?: string;
+  spreadPrice: number;
+};
